Add dsl tests

+++ predef/src/main/scala/io/taig/gandalf/predef/NumericComparison.scala
@@ -4,7 +4,9 @@ import io.taig.gandalf.core.{ Rule, Validation }
 
 import scala.language.higherKinds
 
-abstract class NumericComparison[R[_] <: Rule.Condition]( f: ( Double, Double ) ⇒ Boolean ) {
+abstract class NumericComparison[R[_] <: Rule.Condition](
+        f: ( Double, Double ) ⇒ Boolean
+) {
     @inline
     private def lhs[U: Numeric]( input: U ): Double =
         implicitly[Numeric[U]].toDouble( input )
@@ -19,10 +21,10 @@ abstract class NumericComparison[R[_] <: Rule.Condition]( f: ( Double, Double ) 
         Validation.condition( input ⇒ f( lhs( input ), valueOf[T] ) )
 
     implicit def int[T <: Int: ValueOf, U: Numeric]: Validation[R[T], U, U] =
-        Validation.condition( input ⇒ f( lhs( input ), valueOf[T].toDouble ) )
+        Validation.condition( input ⇒ f( lhs( input ), valueOf[T] ) )
 
     implicit def long[T <: Long: ValueOf, U: Numeric]: Validation[R[T], U, U] =
-        Validation.condition( input ⇒ f( lhs( input ), valueOf[T].toDouble ) )
+        Validation.condition( input ⇒ f( lhs( input ), valueOf[T] ) )
 
     implicit def short[T <: Short: ValueOf, U: Numeric]: Validation[R[T], U, U] =
         Validation.condition( input ⇒ f( lhs( input ), valueOf[T] ) )
